requirements.txt
aws-cdk-lib==2.120.0
constructs>=10.0.0,<11.0.0
pydantic>=2.0.0,<3.0.0
boto3>=1.34.0
aws-lambda-powertools>=2.30.0